+++ src/components/news-item-list/index.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import { NewsItem } from '../';
-
-
-
+import * as ducks from '../../ducks';
 
 export const NewsItemList = ( {ids} ) =>  {
     return(
