+++ src/ducks/data/index.js
@@ -3,7 +3,7 @@ import itemsIds from "./itemsIds";
 import items from "./items";
 import { mapObj, augmentSelectorWith } from "../../utils";
 
-export const ns = "data";
+export const ns = "data-duck";
 
 const root = state => state[ns];
 export const selectors = {
