+++ src/ducks/index.js
@@ -1,1 +1,2 @@
 export { default as ui } from './ui';
+export { default as data } from './data';
