+++ src/index.js
@@ -12,8 +12,9 @@ import { combineReducers } from 'redux';
 
 
 const rootReducer = combineReducers({
-    ...ducks.ui.Reducer, // { ui: ui.rawReducer }
-    data: dataReducer, // { example: example.rawReducer }
+    ...ducks.ui.reducer, // { ui: ui.rawReducer }
+    ...ducks.data.reducer, // { example: example.rawReducer }
+    data: dataReducer
 }); // combineReducers({ ui: ui.rawReducer, example: example.rawReducer })
